Pointers are 32-bit
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@4919 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/win/gl_draw.h
@@ -51,7 +51,7 @@ extern uint8_t En2xSaI;
 extern uint8_t sl_intensity;
 extern uint8_t FilteredGUI;
 extern uint8_t GUIOn2;
-extern uint8_t vidbuffer;
+extern uint32_t vidbuffer;
 extern uint8_t curblank;
 extern uint8_t GUIRESIZE[];
